fixup! chore: Check OAuth token when workspace starts
