+++ packages/dashboard-frontend/src/store/FactoryResolver/index.ts
@@ -27,6 +27,7 @@ import normalizeDevfileV1 from './normalizeDevfileV1';
 import { selectDefaultNamespace } from '../InfrastructureNamespaces/selectors';
 import { getYamlResolver } from '../../services/dashboard-backend-client/yamlResolverApi';
 import { DEFAULT_REGISTRY } from '../DevfileRegistries';
+import { isOAuthResponse } from '../../services/oauth';
 
 const WorkspaceClient = container.get(CheWorkspaceClient);
 
@@ -40,15 +41,6 @@ export type OAuthResponse = {
   message: string;
 };
 
-export function isOAuthResponse(responseData: any): responseData is OAuthResponse {
-  if (
-    responseData?.attributes?.oauth_provider &&
-    responseData?.attributes?.oauth_authentication_url
-  ) {
-    return true;
-  }
-  return false;
-}
 export interface ResolverState extends FactoryResolver {
   optionalFilesContent?: {
     [fileName: string]: string;
